+++ core/market/decentralized/Cargo.toml
@@ -10,6 +10,7 @@ market-test-suite = []
 [dependencies]
 ya-client = "0.3"
 ya-core-model = { version = "0.1", features = ["market", "net"] }
+ya-net = "0.1"
 ya-persistence = "0.2"
 ya-service-api = "0.1"
 ya-service-api-interfaces = "0.1"
