+++ core/model/src/net.rs
@@ -8,8 +8,7 @@ pub const PUBLIC_PREFIX: &str = "/public";
 pub mod local {
     use serde::de::DeserializeOwned;
     use serde::{Deserialize, Serialize};
-    use std::future::Future;
-    use ya_service_bus::{typed as bus, Handle, RpcEndpoint, RpcMessage};
+    use ya_service_bus::RpcMessage;
 
     pub const BUS_ID: &str = "/local/net";
 
@@ -102,38 +101,5 @@ pub mod local {
         SubscribeError(#[from] SubscribeError),
         #[error(transparent)]
         GsbError(#[from] ya_service_bus::error::Error),
-    }
-
-    pub async fn bind_broadcast_with_caller<MsgType, Output, F>(
-        broadcast_address: &str,
-        handler: F,
-    ) -> Result<Handle, BindBroadcastError>
-    where
-        MsgType: BroadcastMessage + Send + Sync + 'static,
-        Output: Future<
-                Output = Result<
-                    <SendBroadcastMessage<MsgType> as RpcMessage>::Item,
-                    <SendBroadcastMessage<MsgType> as RpcMessage>::Error,
-                >,
-            > + 'static,
-        F: FnMut(String, SendBroadcastMessage<MsgType>) -> Output + 'static,
-    {
-        log::debug!("Creating broadcast topic {}.", MsgType::TOPIC);
-
-        // We send Subscribe message to local net, which will create Topic
-        // and add broadcast_address to be endpoint, which will be called, when someone
-        // will broadcast any Message related to this Topic.
-        let subscribe_msg = MsgType::into_subscribe_msg(broadcast_address);
-        bus::service(BUS_ID).send(subscribe_msg).await??;
-
-        log::debug!(
-            "Binding handler '{}' for broadcast topic {}.",
-            broadcast_address,
-            MsgType::TOPIC
-        );
-
-        // We created endpoint address above. Now we must add handler, which will
-        // handle broadcasts forwarded to this address.
-        Ok(bus::bind_with_caller(broadcast_address, handler))
     }
 }
